refactor:  created a lib/shareablity folder and moved relevant files into it.  Updated imports.
