+++ src/lib/shareability/researchBriefPdf.jsx
