+++ src/lib/shareability/researchPanelAccent.js
